pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mmfwd"
version = "0.0.0"
authors = [
  { name="David Timber", email="dxdt@dev.snart.me" },
]
description = "Forward calls and texts using ModemManager framework"
readme = "README.md"
requires-python = ">= 3.6"
dependencies = [ "PyYAML", "PyGObject" ]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/dxdxdt/mmfwd/"
Issues = "https://github.com/dxdxdt/mmfwd/issues"
